# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Gamified Task Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List all tasks
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [todo, in_progress, completed]
        - name: priority
          in: query
          schema:
            type: string
            enum: [critical, high, medium, low]
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [priority, dueDate, vpValue, createdAt]
        - name: projectId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/bulk-reschedule:
    patch:
      operationId: bulkReschedule
      tags: [tasks]
      summary: Move multiple tasks to a new due date
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkRescheduleInput"
      responses:
        "200":
          description: Reschedule result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkRescheduleResult"
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a task by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Not found
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /tasks/{id}/complete:
    post:
      operationId: completeTask
      tags: [tasks]
      summary: Mark a task as complete and award VP
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Completed task with VP award info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskCompleteResult"
...
components:
...
  schemas:
...
    Task:
      type: object
      required: [id, title, status, priority, vpValue, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [todo, in_progress, completed]
        priority:
          type: string
          enum: [critical, high, medium, low]
        vpValue:
          type: integer
        dueDate:
          type: ["string", "null"]
        startDate:
          type: ["string", "null"]
        calendarDate:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        projectId:
          type: ["integer", "null"]
        estimatedMinutes:
          type: ["integer", "null"]
        actualMinutes:
          type: ["integer", "null"]
        links:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/TaskLink"
        notes:
          type: ["string", "null"]
        checklistCount:
          type: integer
        checklistCompleted:
          type: integer
    TaskInput:
      type: object
      required: [title, priority]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        priority:
          type: string
          enum: [critical, high, medium, low]
        dueDate:
          type: string
        startDate:
          type: string
        calendarDate:
          type: string
        vpValue:
          type: integer
        projectId:
          type: integer
        estimatedMinutes:
          type: integer
          minimum: 1
        notes:
          type: string
        links:
          type: array
          items:
            $ref: "#/components/schemas/TaskLink"
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        status:
          type: string
          enum: [todo, in_progress, completed]
        priority:
          type: string
          enum: [critical, high, medium, low]
        dueDate:
          type: string
        startDate:
          type: string
        calendarDate:
          type: string
        vpValue:
          type: integer
        projectId:
          type: integer
        estimatedMinutes:
          type: integer
          minimum: 1
        actualMinutes:
          type: integer
          minimum: 0
        notes:
          type: string
        links:
          type: array
          items:
            $ref: "#/components/schemas/TaskLink"
    TaskLink:
      type: object
      required: [url]
      properties:
        url:
          type: string
        label:
          type: string
    TaskCompleteResult:
      type: object
      required: [task, vpAwarded, multiplier, newTotal, tierUp]
      properties:
        task:
          $ref: "#/components/schemas/Task"
        vpAwarded:
          type: integer
        multiplier:
          type: number
        newTotal:
          type: integer
        tierUp:
          type: boolean
        newTier:
          type: ["integer", "null"]
    BulkRescheduleInput:
      type: object
      required: [taskIds, newDate]
      properties:
        taskIds:
          type: array
          items:
            type: integer
          minItems: 1
        newDate:
          type: string
    BulkRescheduleResult:
      type: object
      required: [updated]
      properties:
        updated:
          type: integer
...