# lib/api-zod/src/generated/types/task.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Gamified Task Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { TaskLink } from './taskLink';
import type { TaskPriority } from './taskPriority';
import type { TaskStatus } from './taskStatus';

export interface Task {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  status: TaskStatus;
  priority: TaskPriority;
  vpValue: number;
  /** @nullable */
  dueDate?: string | null;
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  calendarDate?: string | null;
  /** @nullable */
  completedAt?: string | null;
  createdAt: string;
  /** @nullable */
  projectId?: number | null;
  /** @nullable */
  estimatedMinutes?: number | null;
  /** @nullable */
  actualMinutes?: number | null;
  /** @nullable */
  links?: TaskLink[] | null;
  /** @nullable */
  notes?: string | null;
  checklistCount?: number;
  checklistCompleted?: number;
}