# lib/api-zod/src/generated/types/taskInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Gamified Task Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { TaskInputPriority } from './taskInputPriority';
import type { TaskLink } from './taskLink';

export interface TaskInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  priority: TaskInputPriority;
  dueDate?: string;
  startDate?: string;
  calendarDate?: string;
  vpValue?: number;
  projectId?: number;
  /** @minimum 1 */
  estimatedMinutes?: number;
  notes?: string;
  links?: TaskLink[];
}